Added exchange templating tool into platform services.
