+++ platform/service/exchange_template_files/main_file.tmpl
