+++ platform/service/exchange_template_files/test_file.tmpl
