+++ platform/service/exchange_template_files/type_file.tmpl
